+++ widget/dialog.h
@@ -20,7 +20,7 @@ public:
     double getWidth();
     void setIsClick(bool isclose);                             //关闭串口设置isclick为false
     bool getIsClick();
-    int getRatio() const;
+    double getRatio() const;
 
 private slots:
     void on_btn_ok_clicked();
